integrations/langchain-python/src/langchain_hyperspace/client.py
"""HyperspaceDB Python gRPC client.

This module provides a simple Python client for HyperspaceDB using gRPC.
It will be used by the LangChain integration.
"""

from typing import Dict, List, Optional, Tuple
import grpc


class HyperspaceClient:
    """Simple Python client for HyperspaceDB.
    
    This is a lightweight wrapper around the gRPC API.
    For production use, consider using the full Python SDK.
    """
    
    def __init__(
        self,
        host: str = "localhost",
        port: int = 50051,
        api_key: Optional[str] = None,
    ):
        """Initialize client.
        
        Args:
            host: Server host
            port: Server gRPC port
            api_key: Optional API key for authentication
        """
        self.host = host
        self.port = port
        self.api_key = api_key
        
        # Create channel
        self._channel = grpc.insecure_channel(f"{host}:{port}")
        
        # TODO: Import generated protobuf stubs
        # from langchain_hyperspace.generated import hyperspace_pb2_grpc
        # self._stub = hyperspace_pb2_grpc.DatabaseStub(self._channel)
        
    def _get_metadata(self) -> List[Tuple[str, str]]:
        """Get gRPC metadata with API key."""
        if self.api_key:
            return [("x-api-key", self.api_key)]
        return []
    
    def create_collection(
        self,
        name: str,
        dimension: int,
        metric: str = "l2",
    ) -> bool:
        """Create a new collection.
        
        Args:
            name: Collection name
            dimension: Vector dimension
            metric: Distance metric
            
        Returns:
            True if successful
        """
        # TODO: Implement using protobuf
        # request = hyperspace_pb2.CreateCollectionRequest(
        #     name=name,
        #     dimension=dimension,
        #     metric=metric,
        # )
        # response = self._stub.CreateCollection(request, metadata=self._get_metadata())
        # return response.success
        return True
    
    def insert(
        self,
        collection: str,
        id: int,
        vector: List[float],
        metadata: Dict[str, str],
    ) -> bool:
        """Insert a vector.
        
        Args:
            collection: Collection name
            id: Vector ID
            vector: Vector data
            metadata: Metadata dictionary
            
        Returns:
            True if successful
        """
        # TODO: Implement using protobuf
        # request = hyperspace_pb2.InsertRequest(
        #     collection=collection,
        #     id=id,
        #     vector=vector,
        #     metadata=metadata,
        # )
        # response = self._stub.Insert(request, metadata=self._get_metadata())
        # return response.success
        return True
    
    def search(
        self,
        collection: str,
        vector: List[float],
        k: int = 10,
    ) -> List[Tuple[int, float, Dict[str, str]]]:
        """Search for similar vectors.
        
        Args:
            collection: Collection name
            vector: Query vector
            k: Number of results
            
        Returns:
            List of (id, distance, metadata) tuples
        """
        # TODO: Implement using protobuf
        # request = hyperspace_pb2.SearchRequest(
        #     collection=collection,
        #     vector=vector,
        #     k=k,
        # )
        # response = self._stub.Search(request, metadata=self._get_metadata())
        # return [(r.id, r.distance, dict(r.metadata)) for r in response.results]
        return []
    
    def get_digest(self, collection: str) -> Dict:
        """Get collection digest.
        
        Args:
            collection: Collection name
            
        Returns:
            Digest dictionary
        """
        # TODO: Implement using protobuf
        # request = hyperspace_pb2.DigestRequest(collection=collection)
        # response = self._stub.GetDigest(request, metadata=self._get_metadata())
        # return {
        #     "logical_clock": response.logical_clock,
        #     "state_hash": response.state_hash,
        #     "buckets": list(response.buckets),
        #     "count": response.count,
        # }
        return {
            "logical_clock": 0,
            "state_hash": 0,
            "buckets": [0] * 256,
            "count": 0,
        }
    
    def delete(self, collection: str, id: int) -> bool:
        """Delete a vector.
        
        Args:
            collection: Collection name
            id: Vector ID
            
        Returns:
            True if successful
        """
        # TODO: Implement using protobuf
        # request = hyperspace_pb2.DeleteRequest(
        #     collection=collection,
        #     id=id,
        # )
        # response = self._stub.Delete(request, metadata=self._get_metadata())
        # return response.success
        return True
    
    def close(self) -> None:
        """Close the gRPC channel."""
        if hasattr(self, "_channel"):
            self._channel.close()
    
    def __enter__(self):
        """Context manager entry."""
        return self
    
    def __exit__(self, exc_type, exc_val, exc_tb):
        """Context manager exit."""
        self.close()
